+++ rust-src/bulletproofs/src/utils.rs
@@ -1,0 +1,31 @@
+use crypto_common::*;
+use crypto_common_derive::*;
+use curve_arithmetic::Curve;
+use rand::*;
+
+/// Struct containing generators G and H needed for range proofs
+#[allow(non_snake_case)]
+#[derive(Debug, Clone, Serialize, SerdeBase16Serialize)]
+pub struct Generators<C: Curve> {
+    #[size_length = 4]
+    pub G_H: Vec<(C, C)>,
+}
+
+impl<C: Curve> Generators<C> {
+    /// Generate a list of generators of a given size.
+    pub fn generate(n: usize, csprng: &mut impl Rng) -> Self {
+        let mut gh = Vec::with_capacity(n);
+        for _ in 0..n {
+            let x = C::generate(csprng);
+            let y = C::generate(csprng);
+            gh.push((x, y));
+        }
+        Self { G_H: gh }
+    }
+
+    pub fn take(&self, nm: usize) -> Self {
+        Self {
+            G_H: self.G_H[0..nm].to_vec(),
+        }
+    }
+}
